refactoring tests
